+++ tests/TestCase/FilesystemTest.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 
 namespace Tools\Test;
 
+use InvalidArgumentException;
 use Symfony\Component\Filesystem\Exception\IOException;
 use Symfony\Component\Finder\Exception\DirectoryNotFoundException;
 use Tools\Filesystem;
@@ -49,6 +50,18 @@ class FilesystemTest extends TestCase
         $expected = DS . 'tmp' . DS;
         $this->assertSame($expected, $this->Filesystem->addSlashTerm(DS . 'tmp'));
         $this->assertSame($expected, $this->Filesystem->addSlashTerm($expected));
+    }
+
+    /**
+     * Test for `concatenate()` method
+     * @test
+     */
+    public function testConcatenate()
+    {
+        $this->assertSame('dir', $this->Filesystem->concatenate('dir'));
+        $this->assertSame('dir' . DS . 'subdir', $this->Filesystem->concatenate('dir', 'subdir'));
+        $this->assertSame('dir' . DS . 'subdir', $this->Filesystem->concatenate('dir' . DS, 'subdir'));
+        $this->assertSame('dir' . DS . 'subdir' . DS . 'subsubdir', $this->Filesystem->concatenate('dir', 'subdir', 'subsubdir'));
     }
 
     /**
@@ -198,6 +211,20 @@ class FilesystemTest extends TestCase
         //Resets the ROOT value, removing the final slash
         putenv('ROOT=' . rtrim(ROOT, DS));
         $this->assertSame(rtrim(ROOT, DS), $this->Filesystem->getRoot());
+    }
+
+    /**
+     * Test for `makePathAbsolute()` method
+     * @test
+     */
+    public function testMakePathAbsolute()
+    {
+        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute(TMP . 'dir', TMP));
+        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute('dir', TMP));
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The start path `relativePath` is not absolute');
+        $this->Filesystem->makePathAbsolute('dir', 'relativePath');
     }
 
     /**