+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * all `_or_fail()` methods have been removed.
 
 ## 1.4 branch
+### 1.4.5
+* added `Filesystem::concatenate()` method;
+* added `Filesystem::makePathAbsolute()` method.
+
 ### 1.4.4
 * added `Filesystem` class, all filesystem global function are now deprecated;
 * added `Filesystem::getRoot()` method;